Moved TableFunctions to a new 'helpers' module.
